Pack browser compacting stuff (and extracted card to its own file)
